improve logs
